API update: new `throws` option available
