--- packages/plexus-core/src/api.ts
@@ -1,2 +1,0 @@
-export { api, PlexusApi, PlexusApiConfig, PlexusApiRes, ApiInstance } from '@plexusjs/api'
-// export { api, PlexusApi, PlexusApiConfig, PlexusApiRes }
